requirements.txt
# python3.12 project
behave==1.2.6
python-on-whales==0.74.0
PyHamcrest==2.1.0
deepdiff==8.0.1
pymongo==4.10.1
psycopg2-binary==2.9.10
SQLAlchemy==2.0.36
PyMySQL==1.1.1
allure-behave==2.13.5
pytz==2024.2
xmldiff==2.7.0
jq==1.8.0
